phong conversion couleur fix
