+++ scenes/ref2.rt
@@ -9,33 +9,29 @@ C	0,1,30		0,0,1			70
 
 #	Light
 #	position	intensity		
-L	3.5,-5,-10		1
-sp	3.5,-6,-10		0.5  		     	255,0,255
-sp	3.5,-6,-30		2 	 		     	0,50,255
-sp	3.5,-6,100		10 	 		     	0,50,255
-
+L	-3,-8,-10		1
 
 #	Sphere
 #	center		diameter		color
 # violet
-sp	0,0,-10		1.5  		     	255,0,255
+sp	3.5,0,-10		1.5  		     	255,0,255
 
 # ROUGE
 sp	2,0,-10		1.5  		     	255,33,0
 
 # VERTE
-sp	4,-3,-10	1  		     	0,255,0
+sp	-4,-3,-10	1  		     	0,255,0
 
 # grand sphere blue
-sp	0,10,-10	18				0,255,255
+sp	0,7,-10	12				0,255,255
 
 #	Plan
 #	point		normal			color
-pl	6,0,0			-1,0,0		0,0,43
-pl 	0,0,-30		0,0,-1			40,50,225 
-pl	-5,0,0			1,0,0		255,255,0
+# pl	6,0,0		-1,0,0			50,50,25
+# pl 	0,0,-30		0,0,-1			0,0,225 
 # pl 	0,0,-30		0,0,-1			0,0,0 
 # pl 	0,-1,0		0,1,0			0,255,0 
+# pl	-5,0,0			1,0,0		255,255,0
 # pl	6,0,0			-1,0,0		255,32,43
