+++ scenes/scene1.rt
@@ -1,0 +1,35 @@
+A 0.2 255,255,255
+
+C 0,30,20 0,0,-1 90
+L 0,30,-70 0.4
+
+sp -30,0,-70 8  150,20,0
+sp 0,0,-70 8  150,20,0
+sp 30,0,-70 8  150,20,0
+sp -30,30,-70 8  150,20,0
+sp 30,30,-70 8  150,20,0
+sp -30,60,-70 8  150,20,0
+sp 0,60,-70 8  150,20,0
+sp 30,60,-70 8  150,20,0
+
+sp -30,0,-40 8  150,20,0
+sp 0,0,-40 8  150,20,0
+sp 30,0,-40 8  150,20,0
+sp -30,30,-40 8  150,20,0
+sp 30,30,-40 8  150,20,0
+sp -30,60,-40 8  150,20,0
+sp 0,60,-40 8  150,20,0
+sp 30,60,-40 8  150,20,0
+
+sp -30,0,-100 8  150,20,0
+sp 0,0,-100 8  150,20,0
+sp 30,0,-100 8  150,20,0
+sp -30,30,-100 8  150,20,0
+sp 0,30,-100 8  150,150,150
+sp 30,30,-100 8  150,20,0
+sp -30,60,-100 8  150,20,0
+sp 0,60,-100 8  150,20,0
+sp 30,60,-100 8  150,20,0
+
+cy 0,30,-70 0,1,0 110 90 0,20,150
+
