+++ scenes/scene2.rt
@@ -1,0 +1,33 @@
+A 0.4 255,255,255
+
+C 0,5,0 0,0,-1 90
+L 0,30,0 0.5
+
+cy -35,15,-40 0,-0.3,-1 2 40 30,50,10
+cy -25,15,-40 0,-0.3,-1 2 40 30,50,10
+cy -15,15,-40 0,-0.3,-1 2 40 30,50,10
+cy -5,15,-40 0,-0.3,-1 2 40 30,50,10
+cy 5,15,-40 0,-0.3,-1 2 40 30,50,10
+cy 15,15,-40 0,-0.3,-1 2 40 30,50,10
+cy 25,15,-40 0,-0.3,-1 2 40 30,50,10
+cy 35,15,-40 0,-0.3,-1 2 40 30,50,10
+
+cy -20,5,-40 0,0,-1 2 30 30,150,0
+cy -15,5,-40 0,0,-1 2 30 30,150,0
+cy -10,5,-40 0,0,-1 2 30 30,150,0
+cy -5,5,-40 0,0,-1 2 30 30,150,0
+cy 5,5,-40 0,0,-1 2 30 30,150,0
+cy 10,5,-40 0,0,-1 2 30 30,150,0
+cy 15,5,-40 0,0,-1 2 30 30,150,0
+cy 20,5,-40 0,0,-1 2 30 30,150,0
+
+cy -20,-5,-40 0,0.3,-1 2 20 10,80,20
+cy -15,-5,-40 0,0.3,-1 2 20 10,80,20
+cy -10,-5,-40 0,0.3,-1 2 20 10,80,20
+cy -5,-5,-40 0,0.3,-1 2 20 10,80,20
+cy 5,-5,-40 0,0.3,-1 2 20 10,80,20
+cy 10,-5,-40 0,0.3,-1 2 20 10,80,20
+cy 15,-5,-40 0,0.3,-1 2 20 10,80,20
+cy 20,-5,-40 0,0.3,-1 2 20 10,80,20
+
+pl 0,0,-40 0,0,-1 100,0,70
