+++ scenes/scene3.rt
@@ -1,0 +1,9 @@
+A 0.3 255,255,255
+
+C 0,0,0 0,0,-1 70
+L 10,40,10 0.7
+
+cy 10,0,-40 0,1,0 3 40 70,150,20
+pl 0,-20,0 0,1,0 100,0,70
+pl 0,0,-50 0,0,-1 40,20,180
+sp -10,0,-40 10  150,20,0
