+++ scenes/scene4.rt
@@ -1,0 +1,29 @@
+A 0.2 255,255,255
+
+C 0,5,0 0,0,-1 90
+L 0,30,20 0.5
+
+cy 30,0,-45 0,1,0 6 100 120,15,15
+cy 25,0,-40 0,1,0 5 100 120,15,15
+cy 15,0,-40 0,1,0 7 100 120,15,15
+cy 10,0,-40 0,1,0 5 100 120,15,15
+cy 5,0,-40 0,1,0 9 100 120,15,15
+cy -5,0,-40 0,1,0 5 100 120,15,15
+cy -20,0,-50 0,1,0 10 100 120,15,15
+cy -15,0,-40 0,1,0 5 100 120,15,15
+cy -25,0,-40 0,1,0 7 100 120,15,15
+
+sp 0,5,-15 0.8 251,224,11
+
+sp -1,5,-16 0.5 150,150,150
+sp 1,5,-16 0.5 150,150,150
+
+sp 1,6,-16 0.5 150,150,150
+sp 0,6,-16 0.5 150,150,150
+sp -1,6,-16 0.5 150,150,150
+
+sp -1,4,-16 0.5 150,150,150
+sp 0,4,-16 0.5 150,150,150
+sp 1,4,-16 0.5 150,150,150
+
+pl 0,0,-40 0,0,-1 120,15,15
